+++ app/guid-node/drafts/route.ts
@@ -1,0 +1,7 @@
+import Route from '@ember/routing/route';
+
+export default class GuidNodeDrafts extends Route {
+    model(params: { draftId: string }) {
+        this.transitionTo('registries.drafts.draft-shim', params.draftId);
+    }
+}